Removed unnecessary files from web viewer.

+++ expansion_gl/web/pack.bash
@@ -8,19 +8,22 @@ then
 	exit 1
 fi
 
-function package_data_file
-{
-	local INFILE=$1
-	local BASENAME="$(basename "$INFILE" | cut -f 1 -d '.')"
-	if ! cmp -s "$INFILE" "${BASENAME}.data"
-	then
-		rm -f "${BASENAME}.data" "${BASENAME}.js"
-		python "${EMSDK}/upstream/emscripten/tools/file_packager.py" \
-		  ./${BASENAME}.data \
-		  --preload ${INFILE}@/ \
-		  --js-output=./${BASENAME}.js
-	fi
-}
+INFILE="$1"
 
-package_data_file ./2zsk.pdb
+if [ ! -s "$INFILE" ]
+then
+	echo "Error: no input file '$INFILE'."
+	exit 1
+fi
+
+BASENAME="$(basename "$INFILE" | cut -f 1 -d '.')"
+
+if ! cmp -s "$INFILE" "${BASENAME}.data"
+then
+	rm -f "${BASENAME}.data" "${BASENAME}.js"
+	python "${EMSDK}/upstream/emscripten/tools/file_packager.py" \
+	  ./${BASENAME}.data \
+	  --preload ${INFILE}@/ \
+	  --js-output=./${BASENAME}.js
+fi
